Improve performance for the tileset rendering
As we were using entirely 1x1 tile for the background a really large
grid, so for example, 100 x 100 tiles means we render out 10,000 tile
objects. This causes big slowdowns.

Now instead we use 13x13 tiles for the body and we use elongated tiles
for borders. This means for the same 100 x 100 tiles, rather than having
10 000 objects, we now render out about 100. This means 100x LESS
objects, so performance will dramatically improve.